my_app/requirements.txt
Flask>=2.2
requests>=2.28.0
beautifulsoup4>=4.11.0
PyPDF2>=3.0.0
python-dotenv>=0.21.0